Enhance package-parser
