+++ Sources/package-parser/main.swift
@@ -1,0 +1,15 @@
+/*
+ This source file is part of the Swift.org open source project
+
+ Copyright (c) 2021 Apple Inc. and the Swift project authors
+ Licensed under Apache License v2.0 with Runtime Library Exception
+
+ See http://swift.org/LICENSE.txt for license information
+ See http://swift.org/CONTRIBUTORS.txt for Swift project authors
+ 
+ FIXME: This is a temporary alternative of the frontend implementation.
+*/
+
+import ScriptParse
+
+ScriptParse.main()
